Freeze kinda

+++ Assets/Scripts/Player/Skills/DamageArea.cs
@@ -3,10 +3,10 @@ using UnityEngine;
 
 public class DamageArea : MonoBehaviour
 {
-    public float damage { get; private set; }
+    public int damage { get; private set; }
     public float frequency { get; private set; }
     
-    public void Init(float damage, float timeToVanish, float frequency)
+    public void Init(int damage, float timeToVanish, float frequency)
     {
         this.damage = damage;
         this.frequency = frequency;
